+++ mobile/android/android-components/components/browser/engine-gecko-beta/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngine.kt
@@ -169,7 +169,10 @@ class GeckoEngine(
             set(value) {
                 value?.let { policy ->
                     val activateStrictSocialTracking =
-                        policy.trackingCategories.contains(TrackingProtectionPolicy.TrackingCategory.STRICT)
+                        policy.strictSocialTrackingProtection ?: policy.trackingCategories.contains(
+                            TrackingProtectionPolicy.TrackingCategory.STRICT
+                        )
+
                     runtime.settings.contentBlocking.setStrictSocialTrackingProtection(
                         activateStrictSocialTracking
                     )